requirements.txt
gunicorn==23.0.0
packaging==25.0
parse==1.20.2
WebOb==1.8.9
pytest==9.0.2
requests==2.32.5
requests-wsgi-adapter==0.4.1
pytest-cov==7.0.0
